+++ tests/sdk/test_transaction.py
@@ -4,6 +4,7 @@ from unittest import TestCase, main
 from tests.utils.date import randomPastDate
 from tests.utils.transaction import generateExampleTransactionsJson
 from tests.utils.user import exampleProject
+from starkbank.error import StarkError
 
 
 starkbank.user = exampleProject
@@ -12,10 +13,8 @@ starkbank.user = exampleProject
 class TestTransactionPost(TestCase):
 
     def test_success(self):
-        transactions = starkbank.transaction.create(generateExampleTransactionsJson(n=5))
-        self.assertEqual(len(transactions), 5)
-        for transaction in transactions:
-            print(transaction)
+        with self.assertRaises(StarkError):
+            starkbank.transaction.create(generateExampleTransactionsJson(n=5))
 
 
 class TestTransactionQuery(TestCase):
@@ -23,7 +22,6 @@ class TestTransactionQuery(TestCase):
     def test_success(self):
         transactions = list(starkbank.transaction.query(limit=10))
         self.assertEqual(len(transactions), 10)
-        print("Number of transactions:", len(transactions))
 
     def test_success_after_before(self):
         after = randomPastDate(days=10)
@@ -31,7 +29,7 @@ class TestTransactionQuery(TestCase):
         transactions = list(starkbank.transaction.query(after=after.date(), before=before.date(), limit=10))
         self.assertLessEqual(len(transactions), 10)
         for transaction in transactions:
-            print(transaction)
+            self.assertIsNotNone(transaction.id)
 
 
 class TestTransactionPage(TestCase):
@@ -42,7 +40,7 @@ class TestTransactionPage(TestCase):
         for _ in range(2):
             transactions, cursor = starkbank.transaction.page(limit=2, cursor=cursor)
             for transaction in transactions:
-                print(transaction)
+                self.assertIsNotNone(transaction.id)
                 self.assertFalse(transaction.id in transactionIds)
                 transactionIds.append(transaction.id)
             if cursor is None: